GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - main (refs/gitbutler/main)
   - crates/precompile/src/lib.rs
   - crates/precompile/src/blake2.rs
   - crates/primitives/src/bytecode.rs
   - crates/primitives/src/specification.rs
   - crates/primitives/src/env.rs
   - crates/primitives/src/db.rs
   - crates/primitives/Cargo.toml
   - Cargo.lock
   - crates/interpreter/src/instructions/memory.rs
   - crates/interpreter/src/instructions/stack.rs
   - crates/interpreter/src/interpreter/stack.rs
   - crates/interpreter/src/instructions/i256.rs
   - crates/interpreter/src/instructions/bitwise.rs
   - crates/interpreter/src/inner_models.rs
   - crates/interpreter/src/host.rs
   - crates/interpreter/src/gas.rs
   - crates/interpreter/src/interpreter.rs
   - crates/interpreter/src/lib.rs
   - crates/revm/src/handler.rs
   - crates/revm/src/builder.rs

Your previous branch was: refs/heads/main

The sha for that commit was: 1ad6387d584fcc861378af401714635a5ab5ab83

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch